+++ samples/gesture/AlbumDemo.dart
@@ -71,7 +71,7 @@ class AlbumDemo extends Activity {
       // photo source link
       String link = "$flickerPrefix${authorURLs[i]}/${imgURLs[i]}";
       TextView caption = 
-          new TextView.html("By <a href='$link'>${authors[i]}</a> @ Flickr");
+          new TextView.fromHTML("By <a href='$link'>${authors[i]}</a> @ Flickr");
       caption.classes.add("photo-caption");
       caption.style.userSelect = "none";
       caption.profile.text = "location: south center";