+++ samples/issues/Issue3.dart
@@ -29,7 +29,7 @@ class Issue3 extends Activity {
     _left.on.click.add((event) {_left.visible = false; _right.visible = true; mainView.requestLayout();});
     _right.on.click.add((event) {_left.visible = true; _right.visible = false; mainView.requestLayout();});
 
-    mainView.addChild(new TextView.html('''
+    mainView.addChild(new TextView.fromHTML('''
         <ul><li>Click FIRST TEXT and you shall see it disappears and SECOND TEXT shows up</li>
         <li>Click SECOND TEXT and you shall see the reversed response</li></ul>
         '''));