+++ samples/test/TestViewTag.dart
@@ -7,10 +7,10 @@ class TestViewTag extends Activity {
 
   void onCreate_() {
     /* Notice that this is only for testing purpose. This sample is better to be
-       done by use of the TextView.html constructor as follows (not exactly
+       done by use of the TextView.fromHTML constructor as follows (not exactly
        equalivalent but you got the point).
 
-       new TextView.html('''
+       new TextView.fromHTML('''
        <dl>
         <li type="i" value="$index">This is the <b>$index</b> item</li>
        </dl>